3/10 update

+++ src/main/java/com/example/dreamtree/service/MemberService.java
@@ -5,6 +5,7 @@ import com.example.dreamtree.model.Member;
 import com.example.dreamtree.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
@@ -15,8 +16,12 @@ public class MemberService {
 
     private final MemberRepository repository;
 
-    public Optional<Member> selectMember(String userId) {
-        return repository.findMemberByUserId(userId);
+    public Member selectMemberById(String userId) {
+        return repository.findMemberByUserId(userId).get();
+    }
+
+    public Member selectMemberByNick(String userNick) {
+        return repository.findMemberByUserNick(userNick).get();
     }
 
     public Member createMember(MemberDto member) {
@@ -32,12 +37,26 @@ public class MemberService {
         return repository.save(member1);
     }
 
-    public Member updateMember(Member member) {
-        return repository.save(member);
+    @Transactional
+    public boolean updateMember(String userId, Member member) {
+        Member selected = repository.findMemberByUserId(userId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 회원이 없습니다." + userId));
+        if (selected != null) {
+            repository.save(member);
+            return true;
+        }
+        return false;
     }
 
-    public void deleteMember(Member member) {
 
-        repository.delete(member);
+    public String deleteMember(String userId) {
+        Member member = repository.findMemberByUserId(userId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 회원이 없습니다." + userId));
+
+        if (member !=null) {
+            repository.delete(member);
+            return "회원이 삭제되었습니다.";
+        }
+        return "해당 회원이 없습니다.";
     }
 }
